Fixing the format of the 'side' variable in NewOrder
The manual described the format of the variable 'side' in the function
'NewOrder' as "Buy" or "Sell".

Oanda requires this variable to be all lower case.

This change fixes the manual so that it is all lower case and adds a
line of code to the NewOrder function so that it will run regardless of
capitalization errors.