Complete Overhaul of function output format, Oanda errors and IDs - OAPI Human
This update changes the way function outputs are handled, in previous
versions many functions ouput cell arrays of structs, this was the way
MatLab interpreted the raw data. Since then I have decided that
outputting large sets of data as single structs is much better, now
functions such as GetHistory and GetTransactionHistory display like
large spreadsheets.

Oanda errors no longer cause critical stops.
In previous versions many functions would return an error if Oanda
responsed with an error message, now Oanda errors no longer cause matlab
errors and the functions continue to operate, printing the error details
and successfully completing by returning the error messages to the
answer.

Many functions in previous versions output IDs such as accountId and
transaction IDs as numbers, this is not a good format for this data as
it would be non-sensical to use them in any arithmetic operation and
OANDA will only accept them as input if they are in the form of strings,
all IDs are now converted to strings before functions are completed.